Use grid and stack and reduce code
